Add files via upload
A Measure of Optimal Foreign Currency Reserves in Jamaica: I presented this paper at the Annual National Southwestern Economics Association conference, San Antonio, TX in April 2006. This study aims to compute the optimal level of foreign currency reserve holdings of the Bank of Jamaica. A logit model consisting of panel data for the Caribbean and Latin American countries was used to calculate the probability of a crisis