Add files via upload
The Nexus Between Growth and Tax Structure in Jamaica
I collaborated to write this paper which examines the impact of the tax burden and the general tax structure on economic growth in Jamaica for the period 1989- 2006. After controlling for various determinants of growth, our results indicate that the CIT and PAYE had no impact on economic growth, while the effect of the GCT was positive albeit fragile.