Project Management Plan
This project management plan was developed as a requirement for completion of a Certificate in Applied Project Management by Boston University. The entire curriculum was based on the PMI Project Management Body of Knowledge (PMBOK®). 

The project management plan was developed within a team consisting of 5 members (including myself). Some information has been removed to protect the privacy of the group members.