Presented this econometric paper-logit model (pdf)
I presented this paper at the Annual National Southwestern Economics Association conference, San Antonio, TX in April 2006. This study aims to compute the optimal level of foreign currency reserve holdings of the Bank of Jamaica. A logit model consisting of panel data for the Caribbean and Latin America countries was used to calculate the probability of a crisis. Results indicate that a lower reserve to short-term debt ratio is significant in reducing the probability of a currency crisis. The probability estimate was then utilized in calculating the level of optimal reserves. A cost-benefit approach was taken, where the monetary authorities minimize a loss function subject to a wealth constraint. Based on the estimates derived, it can be concluded that the Bank of Jamaica is holding less reserves than is necessary to curtail a currency crisis with an assumed cost of 40.0 per cent of GDP.