Collaborated on this econometric paper (pdf)
I collaborated to write this paper which examines the impact of the tax burden and the general tax structure on economic growth in Jamaica for the period 1989- 2006. After controlling for various determinants of growth, our results indicate that the CIT and PAYE had no impact on economic growth, while the effect of the GCT was positive albeit fragile. Our Findings show also that there is no evidence to support the argument that the average tax burden was growth retarding. Nevertheless, the research lends support to the calls for a rebalancing of the Jamaican tax system towards more indirect forms of taxation.